fix(lint): green the integration gate after the M3-01 merge
`mise run lint` is red at the integration head: the programmatic-launch
work (78690bf2) landed without the gate, so its lint debt only surfaced
once both task branches were merged.

- `LoadOptions` and `LaunchContext` each crossed the four-bool line.
  Both are caller-supplied option bags of independent switches, not a
  state machine, so they are grandfathered with a narrow `#[expect]`
  and a reason, the way clippy.toml prescribes, plus a budget row.
- `reasoning_effort.rs` kept its suite inline; moved to the canonical
  `reasoning_effort/tests.rs` sibling the test-layout family requires.
- `load_options_e2e.rs` declared `clippy::unwrap_used` and
  `clippy::panic` expectations that never fire in a test target;
  removed rather than allowed, and the budgets shrink back to measured.
- Raised the three budgets that genuinely grew (`disallowed_methods@jackin`,
  `rust-function-complexity/jackin-runtime`) and regenerated the
  telemetry error-ownership census for the moved line.

No product behavior changes.
