+++ crates/jackin-core/src/reasoning_effort.rs
@@ -83,29 +83,4 @@ impl FromStr for ReasoningEffort {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parses_every_canonical_spelling_case_insensitively() {
-        for effort in ReasoningEffort::ALL {
-            assert_eq!(effort.as_str().parse::<ReasoningEffort>(), Ok(effort));
-            assert_eq!(
-                effort.as_str().to_uppercase().parse::<ReasoningEffort>(),
-                Ok(effort)
-            );
-        }
-    }
-
-    #[test]
-    fn rejects_an_unknown_spelling_and_names_it() {
-        let error = "maximum".parse::<ReasoningEffort>().unwrap_err();
-        assert_eq!(error.input, "maximum");
-        assert!(error.to_string().contains("maximum"));
-    }
-
-    #[test]
-    fn default_is_medium() {
-        assert_eq!(ReasoningEffort::default(), ReasoningEffort::Medium);
-    }
-}
+mod tests;
