+++ crates/jackin-telemetry/registry/error-ownership-census.md
@@ -66,7 +66,7 @@ This file is generated by `cargo xtask telemetry-registry --generate`. It invent
 | `crates/jackin-runtime/src/runtime/host_attach.rs:62` | immediate consumed-error owner | `IoError` |
 | `crates/jackin-runtime/src/runtime/host_attach.rs:1265` | immediate consumed-error owner | `IoError` |
 | `crates/jackin-runtime/src/runtime/host_attach.rs:1269` | immediate consumed-error owner | `IoError` |
-| `crates/jackin-runtime/src/runtime/launch/launch_runtime.rs:1042` | immediate consumed-error owner | `LaunchFailed` |
+| `crates/jackin-runtime/src/runtime/launch/launch_runtime.rs:1051` | immediate consumed-error owner | `LaunchFailed` |
 | `crates/jackin-runtime/src/runtime/launch/launch_slot.rs:60` | immediate consumed-error owner | `LaunchFailed` |
 | `crates/jackin-runtime/src/runtime/launch/load_cleanup.rs:17` | immediate consumed-error owner | `LaunchFailed` |
 | `crates/jackin-runtime/src/usage_relay.rs:383` | immediate consumed-error owner | `RpcError` |
